[ADD] fitness: introduce the use of Membership products
Before this commit, server actions and automations were used to update the pricelist of a partner.
Now, with the use of the new partnership/membership module, this is automated and users can link a grade of membership (e.g. Silver, Gold...) to a product.
This will automatically set/unset the pricelist (associated with this grade) of the partner when the membership starts/is revoked

task-4853044