[FIX] *: fix issues caused by major changes in Odoo standard
In this commit, we fixed the issues that occurred due to major updates in Odoo’s
standard test cases. These changes impacted our industry modules where the
version key was explicitly set to 1.0, which is now the default value. Since
defining it is no longer necessary, it caused the runbot to fail on the master
branch.

Task-5264919